+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Coinbase parser: fixed regex to handle whole numbers.
 - Conversion tool: handle .xlsx files with incorrect dimensions. ([#342](https://github.com/BittyTax/BittyTax/issues/342))
 - Accounting tool: handle .xlsx files with incorrect dimensions. ([#342](https://github.com/BittyTax/BittyTax/issues/342))
+- KuCoin parser: use UTC offset from Time header for timestamp.
 ### Added
 - Accounting tool: new PDF report format.
 - Config: added classic_report parameter to switch to legacy PDF report format.